+++ src/main/kotlin/id/walt/vclib/registry/VcTypeRegistry.kt
@@ -41,6 +41,7 @@ object VcTypeRegistry {
         register<KybMonoCredential>(KybMonoCredential)
         register<KycCredential>(KycCredential)
         register<ParticipantCredential>(ParticipantCredential)
+        register<LegalPerson>(LegalPerson)
 
         register<PeerReview>(PeerReview)